Fix for IntuitResponse.php
Calling FaultUnit->getIntuitErrorCode() when getting response error would always return 0, because the code attribute in Error element in XML response is not being parsed due to the comparison failed in method setFaultHandler. This is a fix for this issue where it is now calling the getter of the contentType correctly and cater for both application/xml and text/xml content type.