Rename pageNumber to startOffset
Page Number was wildly inaccurate.  Unless your page size was set to 1, Page Number wasn't acting as a page number but instead was acting as a flat offset.  In an environment where you're "paging" through and incrementing by 1, you wind up with 499 duplicate items and 1 new item.

If you want to stick with pageNumber, please multiply the pageNumber by the pageSize when inserting into the query after startPosition.